+++ packages/sharing-editor/public/samples/011_component_gallery/pages/charts.area_chart2.py
@@ -1,16 +1,18 @@
-import numpy as np
 import pandas as pd
 import streamlit as st
+from numpy.random import default_rng as rng
 
-
-@st.cache_data
-def load_data():
-    df = pd.DataFrame(np.random.randn(20, 3), columns=["col1", "col2", "col3"])
-    return df
-
-
-chart_data = load_data()
+df = pd.DataFrame(
+    {
+        "col1": list(range(20)),
+        "col2": rng(0).standard_normal(20),
+        "col3": rng(1).standard_normal(20),
+    }
+)
 
 st.area_chart(
-    chart_data, x="col1", y=["col2", "col3"], color=["#FF0000", "#0000FF"]  # Optional
+    df,
+    x="col1",
+    y=["col2", "col3"],
+    color=["#FF000080", "#0000FF80"],
 )