+++ packages/sharing-editor/public/samples/011_component_gallery/pages/charts.area_chart_steamgraph.py
@@ -1,6 +1,6 @@
 import streamlit as st
 from vega_datasets import data
 
-source = data.unemployment_across_industries()
+df = data.unemployment_across_industries()
 
-st.area_chart(source, x="date", y="count", color="series", stack="center")
+st.area_chart(df, x="date", y="count", color="series", stack="center")