+++ packages/sharing-editor/public/samples/011_component_gallery/pages/charts.bar_chart1.py
@@ -1,20 +1,13 @@
-import numpy as np
 import pandas as pd
 import streamlit as st
+from numpy.random import default_rng as rng
 
+df = pd.DataFrame(
+    {
+        "col1": list(range(20)) * 3,
+        "col2": rng(0).standard_normal(60),
+        "col3": ["a"] * 20 + ["b"] * 20 + ["c"] * 20,
+    }
+)
 
-@st.cache_data
-def load_data():
-    df = pd.DataFrame(
-        {
-            "col1": list(range(20)) * 3,
-            "col2": np.random.randn(60),
-            "col3": ["A"] * 20 + ["B"] * 20 + ["C"] * 20,
-        }
-    )
-    return df
-
-
-chart_data = load_data()
-
-st.bar_chart(chart_data, x="col1", y="col2", color="col3")
+st.bar_chart(df, x="col1", y="col2", color="col3")