+++ packages/sharing-editor/public/samples/011_component_gallery/pages/charts.plotly_chart_event_state.py
@@ -1,7 +1,7 @@
-import streamlit as st
 import plotly.express as px
+import streamlit as st
 
-df = px.data.iris()  # iris is a pandas DataFrame
+df = px.data.iris()
 fig = px.scatter(df, x="sepal_width", y="sepal_length")
 
 event = st.plotly_chart(fig, key="iris", on_select="rerun")