+++ packages/sharing-editor/public/samples/011_component_gallery/pages/layout.container5.py
@@ -1,0 +1,6 @@
+import streamlit as st
+
+flex = st.container(horizontal=True, horizontal_alignment="right")
+
+for card in range(3):
+    flex.button(f"Button {card + 1}")